Fix broken configuration watching
$scope.$watch() only works on single expressions, not arrays.
Instead, use $scope.$watchGroup().  As this functionality was
introduced in Angular.js 1.3, also bump the minimum dependency and
the version used in the demo to that version.  This was found over
in https://github.com/nebulous/infinitude and was causing the dial
to display incorrectly when using an expression for scaleMax that
wasn't ready when the dial was first drawn.